requirements.txt
Flask==3.1.0
flask_socketio==5.4.1
httpx==0.27.2
PyYAML==6.0.2
Requests==2.32.3
user_agents==2.2.0
watchdog==6.0.0
deepgram-sdk>=5.0.0
